requirements.txt
torch
torchvision
onnx
onnxruntime
numpy
pandas
psutil
tqdm
Pillow
streamlit
matplotlib